refactor: share AEAD session implementation across key agreements
Extract common session, nonce, and message-key logic for X25519 and P-256.
Rename the X25519 module explicitly and align public message-key names while
retaining compatibility aliases.
